requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
pyee==8.1.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
slackclient==2.9.3
slackeventsapi==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
yarl==1.6.3
